% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIC.R, R/BIC.R, R/VarCorr.R, R/confint.R,
%   R/deviance.R, R/fitted.R, R/glance.R, R/loglik.R, R/predict.R,
%   R/residuals.R, R/tidy.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{AIC}
\alias{BIC}
\alias{VarCorr}
\alias{confint}
\alias{deviance}
\alias{fitted}
\alias{glance}
\alias{logLik}
\alias{predict}
\alias{residuals}
\alias{tidy}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{generics}{\code{\link[generics]{glance}}, \code{\link[generics]{tidy}}}

  \item{nlme}{\code{\link[nlme]{VarCorr}}}

  \item{stats}{\code{\link[stats]{AIC}}, \code{\link[stats:AIC]{BIC}}, \code{\link[stats]{confint}}, \code{\link[stats]{deviance}}, \code{\link[stats:fitted.values]{fitted}}, \code{\link[stats]{logLik}}, \code{\link[stats]{predict}}, \code{\link[stats]{residuals}}}
}}

