% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{residuals.pmrm_fit}
\alias{residuals.pmrm_fit}
\title{\code{pmrm} residuals.}
\usage{
\method{residuals}{pmrm_fit}(object, ..., data = object$data, adjust = TRUE)
}
\arguments{
\item{object}{A fitted model object of class \code{"pmrm_fit"}
produced by \code{\link[=pmrm_model_decline]{pmrm_model_decline()}} or \code{\link[=pmrm_model_slowing]{pmrm_model_slowing()}}.}

\item{...}{Not used.}

\item{data}{A \code{tibble} or data frame with one row per patient visit.
This is the new data for making predictions.
It must have all the same columns as the original you fit with the model,
except that the outcome column can be entirely absent.
\code{object$data} is an example dataset that will work.
It is just like the original data, except that rows with missing
responses are removed, and the remaining rows are sorted
by patient ID and categorical scheduled visit.}

\item{adjust}{\code{TRUE} or \code{FALSE}.
\code{adjust = TRUE} returns estimates and inference for
covariate-adjusted \code{mu_ij}
values (defined in \code{vignette("models", package = "pmrm")}) for new data.
\code{adjust = FALSE} instead returns inference on \code{mu_ij - W \%*\% gamma},
the non-covariate-adjusted predictions useful in plotting a
continuous disease progression trajectory in \code{\link[=plot.pmrm_fit]{plot.pmrm_fit()}}.}
}
\value{
A numeric vector of residuals corresponding to the
rows of the data supplied in the \code{data} argument.
}
\description{
Compute the residuals (responses minus fitted values)
of a fitted progression model for repeated measures.
}
\examples{
  set.seed(0L)
  simulation <- pmrm_simulate_decline(
    visit_times = seq_len(5L) - 1,
    gamma = c(1, 2)
  )
  fit <- pmrm_model_decline(
    data = simulation,
    outcome = "y",
    time = "t",
    patient = "patient",
    visit = "visit",
    arm = "arm",
    covariates = ~ w_1 + w_2
  )
  str(residuals(fit))
}
\seealso{
Other estimates and predictions: 
\code{\link{VarCorr.pmrm_fit}()},
\code{\link{coef.pmrm_fit}()},
\code{\link{fitted.pmrm_fit}()},
\code{\link{plot.pmrm_fit}()},
\code{\link{pmrm_estimates}()},
\code{\link{pmrm_marginals}()},
\code{\link{predict.pmrm_fit}()},
\code{\link{vcov.pmrm_fit_decline}()},
\code{\link{vcov.pmrm_fit_slowing}()}
}
\concept{estimates and predictions}
