% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confint.R
\name{confint.pmrm_fit}
\alias{confint.pmrm_fit}
\title{Confidence intervals of parameters}
\usage{
\method{confint}{pmrm_fit}(
  object,
  parm = c("theta", "beta", "alpha", "gamma", "sigma", "phi", "rho", "Sigma", "Lambda"),
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{A fitted model object of class \code{"pmrm_fit"}.}

\item{parm}{Character string, name of a family of parameters
to compute confidence intervals.}

\item{level}{Numeric scalar between 0 and 1, confidence level.}

\item{...}{Not used.}
}
\value{
The \code{tibble} returned by \code{\link[=pmrm_estimates]{pmrm_estimates()}}, except
without the columns with point estimates or standard errors.
}
\description{
Compute confidence intervals of the family of model
parameters specified in \code{parm}.
}
\details{
See \code{vignette("models", package = "pmrm")} for details.
}
\examples{
  set.seed(0L)
  simulation <- pmrm_simulate_decline_proportional(
    visit_times = seq_len(5L) - 1,
    gamma = c(1, 2)
  )
  fit <- pmrm_model_decline_proportional(
    data = simulation,
    outcome = "y",
    time = "t",
    patient = "patient",
    visit = "visit",
    arm = "arm",
    covariates = ~ w_1 + w_2
  )
  confint(fit)
}
\seealso{
Other model comparison: 
\code{\link{AIC.pmrm_fit}()},
\code{\link{BIC.pmrm_fit}()},
\code{\link{deviance.pmrm_fit}()},
\code{\link{logLik.pmrm_fit}()},
\code{\link{summary.pmrm_fit}()}
}
\concept{model comparison}
